requirements.txt
python-fasthtml
python-dotenv
openai
sqlalchemy
git+https://github.com/almottier/TapoP100.git@main
wheel
requests
isort
pydantic
webcolors